pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linode-mcp"
version = "0.0.1"
description = "Model Context Protocol server for Linode API"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    {name = "Khaja Omer"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "linode-api4==5.27.0",
    "pydantic>=2.10.6,<3.0.0",
    "python-dotenv==1.0.1",
    "mcp>=1.3.0",
]

[project.urls]
"Homepage" = "https://github.com/komer3/linode-mcp"
"Bug Tracker" = "https://github.com/komer3/linode-mcp/issues"
"Documentation" = "https://github.com/komer3/linode-mcp#readme"

[project.scripts]
linode-mcp = "linode_mcp.server:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py37"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py" 
